dockerfile/docker-easyldap/app/requirements.txt
click==6.6
Flask==0.12
itsdangerous==0.24
Jinja2==2.8.1
MarkupSafe==0.23
python-ldap==2.4.28
Werkzeug==0.11.15
